… .github/workflows/manifold.yml
… CMakeLists.txt
… collider/CMakeLists.txt
… manifold/CMakeLists.txt
… manifold/src/impl.cpp
… manifold/src/manifold.cpp
… manifold/src/properties.cpp
… manifold/src/shared.h
… manifold/src/smoothing.cpp
… manifold/src/sort.cpp
… test/mesh_test.cpp
… utilities/CMakeLists.txt
… utilities/include/par.h
… utilities/include/sparse.h
… utilities/include/vec_dh.h
… utilities/src/detect_cuda.cpp