By reinventing the wheels, we can switch backends dynamically depending on the workload, get better performance and run the build with CUDA on machines without CUDA (switch to OMP automatically).

This patch is large because I have to change every single `thrust::ALGORITHM` to a custom function that determines the backend based on the execution policy, which the execution policy is determined based on the workload size (simply comparing against some constant size). We cannot simply pass a different execution policy to thrust, because the type of different execution policy is different, and the implementation here uses macro to build the functions and a switch to choose the correct invocation.

CUDA GPU detection is done by checking `cudaGetDeviceCount` and setting a global variable. If there is no CUDA GPU devices available, the GPU code path will be skipped and we will only use OpenMP or sequential implementation.

This patch also changed the `VecDH` implementation from `std::vector` + `thrust::universal_vector` to a custom vector implementation, which allows building vectors with uninitialized memory, performs `uninitialized_copy`/`uninitialized_fill` on GPU/CPU based on data size, and also perform memory prefetching to reduce the number of cache misses. This custom vector implementation is required because the old implementation will cause tons of page faults due to the use of `thrust::uninitialized_fill(thrust::device, ...)` when initializing the vector, which slows down the performance a lot.

---

Benchmark:

For CPP and OMP backend, the difference is not very significant. Below is the result for CUDA:
```
Old (#121):
nTri = 512, time = 0.0123526 sec
nTri = 2048, time = 0.0141026 sec
nTri = 8192, time = 0.0195818 sec
nTri = 32768, time = 0.0374303 sec
nTri = 131072, time = 0.120768 sec
nTri = 524288, time = 0.266435 sec
nTri = 2097152, time = 0.924646 sec
nTri = 8388608, time = 4.01982 sec

New:
nTri = 512, time = 0.00276084 sec
nTri = 2048, time = 0.00633459 sec
nTri = 8192, time = 0.00877285 sec
nTri = 32768, time = 0.0524452 sec
nTri = 131072, time = 0.0994011 sec
nTri = 524288, time = 0.285462 sec
nTri = 2097152, time = 0.947242 sec
nTri = 8388608, time = 3.70755 sec
```

Most importantly, the time required for small operation is reduced significantly.

TODO:
- [ ] Run clang-format for the cpp/h files. The indentation is messed up for now, but the diff should be readable.
- [x] Rethink about how we should specify the backend in our build script. Previously the backends are mutually exclusive, but now they are not (we can have OMP, CUDA, CUDA + OMP, CUDA + TBB, etc.).