@pca006132 2022-06-01
I tried running the program with thread sanitizer enabled. On my machine (NixOS), I only got data race warnings for `manifold/src/impl.cpp:644:33`, `manifold/src/impl.cpp:264:24` and `manifold/src/impl.cpp:265:24` which are fine. On docker with Ubuntu however, there are too many data race warnings and it is impossible to identify what causes the nondeterministic behavior. I'm not entirely sure if this bug is caused by my changes or already exists but becomes more apparent due to this change.