@pca006132 2022-06-01
I think this should be ready to merge after fixing the minor issue above (forgot to specify the return type of some functions) and clean up the formatting. For the OMP build on Ubuntu, the Boolean.Gyroid test failed with:
```
[ RUN      ] Boolean.Gyroid
/__w/manifold/manifold/test/mesh_test.cpp:762: Failure
Expected: (result.NumDegenerateTris()) <= (42), actual: 43 vs 42
```

Only with `OMP_NUM_THREADS=2` and quite deterministically failed with this setting.