@pca006132 2022-06-02
Looks like an issue with GitHub API. Quite a lot of outage recently.