@pca006132 2022-06-02
> Also, should the CI just build the limited version of Assimp and avoid the other formats, now that you have that flag? In fact, that could even be the default...

Yes it is now the default.