@elalish 2022-08-27
@pca006132 After a couple days of baffled debug, I finally found the root of a crazy problem that showed up in a test for #187. I thought you might appreciate what I found - I went to bed convincing myself the compiler was broken, but woke up realizing the truth. 

I was seeing manifold errors in the Boolean (printed assertions like `k = 1` and `no intersection`, coming from deep in the internals), which I haven't seen for a long time. It's concerning because this is the guaranteed part of the algorithm - those assertions really shouldn't be possible to hit. Stranger still, it worked fine on the CPU - only CUDA had the error. 

My first guess was that my broad phase had an error so I wasn't checking for all the intersections I should be. I tested this by eliminating it and forcing a brute force check of every intersection. Sure enough, this made CUDA work fine again. But delving into the code, I couldn't find an error anywhere. Finally I got far enough down that I found a case where I was getting two slightly different results from my `Interpolate()` function for exactly the same input (I had to print all the floating-point digits to see the difference in the last decimal place). Sure enough, that violated the assumption of the algorithm and caused the failure, but how could this function not be deterministic?

Finally I remembered: we're using two compilers, one for CPU and one for CUDA. And the choice of which version of each function to use is determined at run time by the vector length. It turns out one pass was long, running CUDA, and one was short, running on the CPU. They must use subtly different operations or ordering that result in slightly different rounding errors occasionally. Overriding the policy to force them onto the same compiled function fixed the issue and verified my hunch. It also explained why my brute force approach seemed to fix the problem: it made all the vectors long enough to choose CUDA.

Anyway, I'm going to change the policy system in #187, so that the policy becomes a member of the Boolean class, evaluated only once or twice and then used consistently throughout the internal operations. I think this may even improve performance a touch, as the real cost of CUDA is in moving the data between the host and device; even short vectors are fast as long as the data is already on the device. This should help us be more consistent.